Use source files instead of bundle
