+++ examples/jsm/csm/Utils.js
@@ -1,0 +1,3 @@
+export function toRad(degrees) {
+	return degrees * Math.PI / 180;
+}
