+++ examples/webgl_cascadedshadowmaps.html
@@ -20,7 +20,7 @@
 
 			import { OrbitControls } from './jsm/controls/OrbitControls.js';
 			import { GUI } from './jsm/libs/dat.gui.module.js';
-			import * as CSM from './js/libs/three-csm.module.js';
+			import * as CSM from './jsm/csm/CSM.js';
 
 			var renderer, scene, camera, controls, csm;
